🚚 rename:	UserSimpleInfo 폴더 추가 (#62)
